+++ javascript/src/components/Idps/idpMapToDataTable.js
@@ -10,7 +10,7 @@ import {useQuery, useQueryClient} from "react-query";
 import {loginsPerCountryKey} from "../../utils/queryKeys";
 import {getLoginsPerCountry} from "../../utils/queries";
 import Spinner from "../Common/spinner";
-import {format} from "date-fns";
+import {formatStartDate, formatEndDate} from "../Common/utils";
 
 const IdpMapToDataTable = ({
                              startDate,
@@ -24,8 +24,8 @@ const IdpMapToDataTable = ({
 
   let params = {
     params: {
-      'startDate': !startDate ? null : format(startDate, "yyyy-MM-dd'T'HH:mm:ss'Z'"),
-      'endDate': !endDate ? null : format(endDate, "yyyy-MM-dd'T'HH:mm:ss'Z'"),
+      'startDate': !startDate ? null : formatStartDate(startDate),
+      'endDate': !endDate ? null : formatEndDate(endDate),
       'tenenv_id': tenenvId,
       'unique_logins': uniqueLogins,
       'idpId': idpId
@@ -45,8 +45,8 @@ const IdpMapToDataTable = ({
   useEffect(() => {
     params = {
       params: {
-        'startDate': !startDate ? null : format(startDate, "yyyy-MM-dd'T'HH:mm:ss'Z'"),
-        'endDate': !endDate ? null : format(endDate, "yyyy-MM-dd'T'HH:mm:ss'Z'"),
+        'startDate': !startDate ? null : formatStartDate(startDate),
+        'endDate': !endDate ? null : formatEndDate(endDate),
         'tenenv_id': tenenvId,
         'unique_logins': uniqueLogins,
         'idpId': idpId