+++ javascript/src/components/Idps/idpsDataTable.js
@@ -16,7 +16,6 @@ import {useCookies} from "react-cookie";
 import {createAnchorElement, formatStartDate, formatEndDate} from "../Common/utils";
 import {toast} from "react-toastify";
 import Spinner from "../Common/spinner";
-import {format} from "date-fns";
 
 const IdpsDataTable = ({
                          spId,
@@ -40,8 +39,8 @@ const IdpsDataTable = ({
 
   let params = {
     params: {
-      'startDate': !startDate ? null : format(startDate, "yyyy-MM-dd'T'HH:mm:ss'Z'"),
-      'endDate': !endDate ? null : format(endDate, "yyyy-MM-dd'T'HH:mm:ss'Z'"),
+      'startDate': !startDate ? null : formatStartDate(startDate),
+      'endDate': !endDate ? null : formatEndDate(endDate),
       'sp': spId,
       'tenenv_id': tenenvId,
       'unique_logins': uniqueLogins
@@ -69,8 +68,8 @@ const IdpsDataTable = ({
   useEffect(() => {
     params = {
       params: {
-        'startDate': !startDate ? null : format(startDate, "yyyy-MM-dd'T'HH:mm:ss'Z'"),
-        'endDate': !endDate ? null : format(endDate, "yyyy-MM-dd'T'HH:mm:ss'Z'"),
+        'startDate': !startDate ? null : formatStartDate(startDate),
+        'endDate': !endDate ? null : formatEndDate(endDate),
         'sp': spId,
         'tenenv_id': tenenvId,
         'unique_logins': uniqueLogins
@@ -110,8 +109,6 @@ const IdpsDataTable = ({
   }, [uniqueLogins, loginsPerIpd.isSuccess && minDateLogins.isSuccess])
 
   const handleStartDateChange = (date) => {
-
-    date = formatStartDate(date);
     if (date != null) {
       setStartDate(date);
     }