+++ javascript/src/components/Users/registeredUsersChart.js
@@ -3,7 +3,8 @@ import {Chart} from "react-google-charts";
 import {
   axisChartOptions,
   convertDateByGroup,
-  getWeekNumber
+  getWeekNumber,
+  parseDateFromISO
 } from "../Common/utils";
 import Select from 'react-select';
 import Row from 'react-bootstrap/Row';
@@ -18,6 +19,7 @@ import {registeredUsersGroupByKey} from "../../utils/queryKeys";
 import {getRegisteredUsersGroupBy} from "../../utils/queries";
 
 const RegisteredUsersChart = ({
+                                showActiveOnly,
                                 tenenvId
                               }) => {
   const [selected, setSelected] = useState(options_group_by[0].value);
@@ -29,7 +31,8 @@ const RegisteredUsersChart = ({
     params: {
       'interval': selected,
       'count_interval': regUsersOptions[selected]["count_interval"],
-      'tenenv_id': tenenvId
+      'tenenv_id': tenenvId,
+      'status': showActiveOnly ? 'A' : null,
     }
   }
 
@@ -43,21 +46,25 @@ const RegisteredUsersChart = ({
   )
 
   useEffect(() => {
-    params = {
-      params: {
-        'interval': selected,
-        'count_interval': regUsersOptions[selected]["count_interval"],
-        'tenenv_id': tenenvId,
+    const fetchData = async () => {
+      params = {
+        params: {
+          'interval': selected,
+          'count_interval': regUsersOptions[selected]["count_interval"],
+          'tenenv_id': tenenvId,
+          'status': showActiveOnly ? 'A' : null,
+        }
+      }
+
+      try {
+        const response = await queryClient.refetchQueries([registeredUsersGroupByKey, {groupBy: selected, params: params}])
+      } catch (error) {
+        // todo: Here we can handle any authentication or authorization errors
+        console.error(RegisteredUsersChart.name + " error: " + error)
       }
     }
-
-    try {
-      const response = queryClient.refetchQueries([registeredUsersGroupByKey, {groupBy: selected, params: params}])
-    } catch (error) {
-      // todo: Here we can handle any authentication or authorization errors
-      console.error(RegisteredUsersChart.name + " error: " + error)
-    }
-  }, [selected, tenenvId])
+    fetchData();
+  }, [selected, tenenvId, showActiveOnly])
 
 
   // Construct the data required for the datatable
@@ -68,8 +75,8 @@ const RegisteredUsersChart = ({
       && !!registeredUsersGroup.data) {
 
       const hticksArray = registeredUsersGroup?.data?.map(element => ({
-          v: new Date(element?.range_date),
-          f: selected === "week" ? getWeekNumber(new Date(element?.range_date)) : new Date(element?.range_date)
+          v: parseDateFromISO(element?.range_date),
+          f: selected === "week" ? getWeekNumber(parseDateFromISO(element?.range_date)) : parseDateFromISO(element?.range_date)
         })
       )
 
@@ -85,9 +92,9 @@ const RegisteredUsersChart = ({
       ]
 
       const charData = registeredUsersGroup?.data?.map(element => ([
-          new Date(element?.range_date),
+          parseDateFromISO(element?.range_date),
           parseInt(element['count']),
-          `<div style="padding:5px 5px 5px 5px;">${convertDateByGroup(new Date(element?.range_date), selected)}<br/>Communities: ${parseInt(element['count'])}</div>`
+          `<div style="padding:5px 5px 5px 5px;">${convertDateByGroup(parseDateFromISO(element?.range_date), selected)}<br/>Users: ${parseInt(element['count'])}</div>`
         ])
       )