+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 ### Changelog
 
+
+## 2026-02-17
+
+### Fixed
+- Remove timezone-aware from all the queries
+
+### Added
+- Checkbox for showing active users only
+
+## 2025-11-20
+
+### Fixed
+- Fix logo url at EGI dev instance
+
+### Added
+- Logins per service per country
+
 ## 2025-06-03
 
 ### Fixed